src/function_app/requirements.txt
azure-functions==1.18.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
numpy==1.26.2
scikit-learn==1.5.2
joblib==1.3.2
